refactor: move formula lookup into workbook layer
